Set proxy configuration at every (re)start of nexus; #22
